backend/requirements.txt
# FastAPI Core
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
aiosqlite>=0.19.0  # For test database (SQLite fallback)

# Data Validation
pydantic>=2.10.0  # Required by openai-agents
pydantic-settings>=2.1.0

# Environment & Config
python-dotenv>=1.0.1

# HTTP Client
httpx>=0.26.0

# Cache & Queue
redis>=5.0.1

# PDF Generation
reportlab>=4.0.9

# AI Generation
openai-agents>=0.1.0  # OpenAI Agents SDK (includes openai dependency)

# Authentication
passlib[bcrypt]>=1.7.4  # Password hashing with bcrypt
python-jose[cryptography]>=3.3.0  # JWT token generation

# Email Service
resend>=2.1.0

# Error Tracking
sentry-sdk[fastapi]>=1.40.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
nest-asyncio>=1.6.0  # For running async migrations in tests